feat: Implement new component QuantitySelector for managing product quantity selection.
Introduces QuantitySelector component to handle quantity selection for a specific product.
Utilizes useState hook to manage the quantity state with an initial value of 1.
Provides functionality to increment and decrement the quantity with incrementQuantity and decrementQuantity functions respectively.
Displays quantity selection UI with buttons (- and +) enclosed in a background container (bg-gray-100).
Updates quantity display dynamically between the buttons.
Includes a call-to-action button (Cta) linked to the product's details page (/products/{product.slug}) for further exploration.